{delete} eliminando las migraciones create-doctor create-usuario create-contacto, excedian las migraciones necesarias y no tenian las asociaciones con las demas tablas ademas create test esta creado con el flujo adecuado de migraciones

+++ AsistenteAED/src/migrations/20210407210014-create-association-personas-doctors.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       d_telefono: {
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER(50)
       },
       d_imagen: {
         type: Sequelize.STRING
